libtysila: added box opcode libtysila: fixed a bug in the register allocator libtysila: added support for two opcode instructions (e.g. x86 ones) that also keep the SSA constraint happy (see tybel.MultipleNode)
git-svn-id: https://www.tysos.org/svn/branches/tysila3@567 21779fe5-bacf-4c11-b3ef-82bd0e1f58e8

+++ JitTestHost/Program.cs
@@ -157,7 +157,7 @@ namespace JitTestHost
             switch (relocation.RelType)
             {
                 case libtysila.x86_64.x86_64_elf64.R_X86_64_64:
-                    int target_addr = symbols[relocation.Name] + relocation.Value;
+                    long target_addr = symbols[relocation.Name] + relocation.Value;
                     reloc = BitConverter.GetBytes((ulong)target_addr);
                     break;
                 default:
