+++ tysila2/Output/Elf64/Elf64_ElfWriter.cs
@@ -296,17 +296,17 @@ namespace Elf64
             ehdr.e_syms.name_to_sym.Add(name, ehdr.e_syms.defined_syms.Count - 1);
         }
 
-        public void AddTextRelocation(int offset, string name, uint rel_type, int value)
+        public void AddTextRelocation(int offset, string name, uint rel_type, long value)
         {
             ehdr.relatext.relocs.Add(new Elf64_Rela_Shdr.Elf64_Rela((ulong)offset, rel_type, value, name));
         }
 
-        public void AddDataRelocation(int offset, string name, uint rel_type, int value)
+        public void AddDataRelocation(int offset, string name, uint rel_type, long value)
         {
             ehdr.reladata.relocs.Add(new Elf64_Rela_Shdr.Elf64_Rela((ulong)offset, rel_type, value, name));
         }
 
-        public void AddRodataRelocation(int offset, string name, uint rel_type, int value)
+        public void AddRodataRelocation(int offset, string name, uint rel_type, long value)
         {
             ehdr.relarodata.relocs.Add(new Elf64_Rela_Shdr.Elf64_Rela((ulong)offset, rel_type, value, name));
         }
